Changes in catalogs structure.
