+++ src/Sylius/Bundle/ReportBundle/Controller/ReportController.php
@@ -13,8 +13,6 @@ namespace Sylius\Bundle\ReportBundle\Controller;
 
 use Sylius\Bundle\ResourceBundle\Controller\ResourceController;
 use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\HttpFoundation\Response;
-use Sylius\Component\Report\Renderer\TableRenderer;
 
 /**
  * @author Mateusz Zalewski <mateusz.zalewski@lakion.com>
